refactor: NavBar 구조 변경 - NavBar 하위 폴더 구조를 정리 - NavBar 컴포넌트의 import 경로 수정
