+++ src/components/ui/navbar/index.ts
@@ -1,0 +1,24 @@
+// 메인 NavBar 컴포넌트
+export { NavBar } from './navbar';
+
+// 복잡한 구조용 컴포넌트들
+export { NavBarComponent } from './navbar-component';
+export { NavBarItem } from './navbar-item';
+export { NavBarDropdown } from './navbar-dropdown';
+export { NavBarDropdownItem } from './navbar-dropdown-item';
+
+// Variants
+export {
+  navbarVariants,
+  navbarItemVariants,
+  navbarDropdownVariants,
+  navbarDropdownItemVariants,
+} from './variants';
+
+// Types
+export type {
+  NavBarProps,
+  NavBarItemProps,
+  NavBarDropdownProps,
+  NavBarDropdownItemProps,
+} from './types';
