+++ src/components/ui/navbar/variants.ts
